Move ClangParser into standard pass.
